Update to make implementation more in-line with LD
